requirements.txt
pyqt6
mysql-connector-python
pandas
openpyxl